// src/components/RestrictedRegion/types.ts
export interface UserMetadata {
    fl: string;
    h: string;
    ip: string;
    ts: string;
    visit_scheme: string;
    uag: string;
    colo: string;
    sliver: string;
    http: string;
    loc: string;
    tls: string;
    sni: string;
    warp: string;
    gateway: string;
    rbi: string;
    kex: string;
  }
  